+++ packages/google_identity_services_web/lib/src/js_interop/package_web_tweaks.dart
@@ -2,34 +2,75 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-/// Provides some useful tweaks to `package:web`.
-library package_web_tweaks;
-
 import 'dart:js_interop';
+
 import 'package:web/web.dart' as web;
+
+// TODO(kevmoo): Make this file unnecessary, https://github.com/dart-lang/web/issues/175
 
 /// This extension gives web.window a nullable getter to the `trustedTypes`
 /// property, which needs to be used to check for feature support.
 extension NullableTrustedTypesGetter on web.Window {
+  /// (Nullable) Bindings to window.trustedTypes.
   ///
+  /// This may be null if the browser doesn't support the Trusted Types API.
+  ///
+  /// See: https://developer.mozilla.org/en-US/docs/Web/API/Trusted_Types_API
   @JS('trustedTypes')
-  external web.TrustedTypePolicyFactory? get nullableTrustedTypes;
-}
+  external TrustedTypePolicyFactory? get nullableTrustedTypes;
 
-/// This extension allows a trusted type policy to create a script URL without
-/// the `args` parameter (which in Chrome currently fails).
-extension CreateScriptUrlWithoutArgs on web.TrustedTypePolicy {
+  /// Bindings to window.trustedTypes.
   ///
-  @JS('createScriptURL')
-  external web.TrustedScriptURL createScriptURLNoArgs(
-    String input,
-  );
+  /// This will crash if accessed in a browser that doesn't support the
+  /// Trusted Types API.
+  ///
+  /// See: https://developer.mozilla.org/en-US/docs/Web/API/Trusted_Types_API
+  @JS('trustedTypes')
+  external TrustedTypePolicyFactory get trustedTypes;
 }
 
 /// This extension allows setting a TrustedScriptURL as the src of a script element,
 /// which currently only accepts a string.
 extension TrustedTypeSrcAttribute on web.HTMLScriptElement {
-  ///
   @JS('src')
-  external set srcTT(web.TrustedScriptURL value);
+  external set trustedSrc(TrustedScriptURL value);
+}
+
+// TODO(kevmoo): drop all of this once `pkg:web` publishes `0.5.1`.
+
+/// Bindings to a JS TrustedScriptURL.
+///
+/// See: https://developer.mozilla.org/en-US/docs/Web/API/TrustedScriptURL
+extension type TrustedScriptURL._(JSObject _) implements JSObject {}
+
+/// Bindings to a JS TrustedTypePolicyFactory.
+///
+/// See: https://developer.mozilla.org/en-US/docs/Web/API/TrustedTypePolicyFactory
+extension type TrustedTypePolicyFactory._(JSObject _) implements JSObject {
+  ///
+  external TrustedTypePolicy createPolicy(
+    String policyName, [
+    TrustedTypePolicyOptions policyOptions,
+  ]);
+}
+
+/// Bindings to a JS TrustedTypePolicy.
+///
+/// See: https://developer.mozilla.org/en-US/docs/Web/API/TrustedTypePolicy
+extension type TrustedTypePolicy._(JSObject _) implements JSObject {
+  ///
+  @JS('createScriptURL')
+  external TrustedScriptURL createScriptURLNoArgs(
+    String input,
+  );
+}
+
+/// Bindings to a JS TrustedTypePolicyOptions (anonymous).
+///
+/// See: https://developer.mozilla.org/en-US/docs/Web/API/TrustedTypePolicyFactory/createPolicy#policyoptions
+extension type TrustedTypePolicyOptions._(JSObject _) implements JSObject {
+  ///
+  external factory TrustedTypePolicyOptions({
+    JSFunction createScriptURL,
+  });
 }