+++ packages/local_auth/local_auth_darwin/darwin/Classes/FLALocalAuthPlugin.h
@@ -6,5 +6,5 @@
 
 #import "messages.g.h"
 
-@interface FLTLocalAuthPlugin : NSObject <FlutterPlugin, FLALocalAuthApi>
+@interface FLALocalAuthPlugin : NSObject <FlutterPlugin, FLADLocalAuthApi>
 @end